+++ lib/sim/simulation-graph-svg/prepare-simulation-graphs/get-string-property.ts
@@ -1,0 +1,8 @@
+export function getStringProperty(
+  value: object,
+  key: string,
+): string | undefined {
+  const propertyValue = (value as Record<string, unknown>)[key]
+  if (typeof propertyValue !== "string") return undefined
+  return propertyValue
+}
