+++ package.json
@@ -25,7 +25,7 @@
     "@vitejs/plugin-react": "5.0.0",
     "biome": "^0.3.3",
     "bun-match-svg": "^0.0.12",
-    "circuit-json": "^0.0.437",
+    "circuit-json": "^0.0.438",
     "esbuild": "^0.20.2",
     "format-si-unit": "^0.0.7",
     "performance-now": "^2.1.0",
